Replace use of `prepend` with `include` in NodeApi
The use of prepend seems harder to understand the order in which modules
receive messages.
