+++ spec/models/qernel/node_api/preset_demand_by_electricity_production_spec.rb
@@ -15,7 +15,7 @@ module Qernel
       node.add_slot(Slot.new(1, nil, Carrier.new(key: :gas), :output).with(conversion: 0.7))
 
       expect { node.query.preset_demand_by_electricity_production = 1000.0 }
-        .to raise_error(/could not find an electricity output/)
+        .to raise_error(/could not find electricity output/)
     end
   end
 end
